refactor: change folder's name
